requirements.txt
biopython>=1.79
requests>=2.28
pandas>=1.5
tqdm>=4.64
python-dotenv>=1.0
